Review ISocialPlatform; make all reviewed code follow the same style

+++ Assets/GooglePlayGames/Runtime/Scripts/Platforms/Android/AndroidClient.cs
@@ -668,7 +668,7 @@ namespace GooglePlayGames.Android
                                     achievement.Id = javaAchievement.Call<string>("getAchievementId");
                                     achievement.Description = javaAchievement.Call<string>("getDescription");
                                     achievement.Name = javaAchievement.Call<string>("getName");
-                                    achievement.Points = javaAchievement.Call<ulong>("getXpValue");
+                                    achievement.Points = (int)javaAchievement.Call<ulong>("getXpValue");
 
                                     long timestamp = javaAchievement.Call<long>("getLastUpdatedTimestamp");
                                     achievement.LastModifiedTime = AndroidJavaConverter.ToDateTime(timestamp);
